Remove legacy Rust venv_bin / venv_shim tooling
