+++ Cargo.toml
@@ -1,11 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
-    "py/tools/py",
-    "py/tools/venv_bin",
-    "py/tools/unpack_bin",
-    "py/tools/venv_shim",
-    "py/tools/runfiles",
+    "py/tools/unpack",
 ]
 
 [workspace.package]
@@ -23,13 +19,10 @@ clap = { version = "4.5.20", features = ["derive"] }
 itertools = "0.13.0"
 miette = { version = "7.2", features = ["fancy"] }
 tempfile = "3.13.0"
-uv-cache = { git = "https://github.com/astral-sh/uv", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
 uv-distribution-filename = { git = "https://github.com/astral-sh/uv.git", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
 uv-extract = { git = "https://github.com/astral-sh/uv.git", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
-uv-install-wheel = { git = "https://github.com/astral-sh/uv", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
+uv-install-wheel = { git = "https://github.com/astral-sh/uv.git", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
 uv-pypi-types = { git = "https://github.com/astral-sh/uv.git", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
-uv-python = { git = "https://github.com/astral-sh/uv", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
-uv-virtualenv = { git = "https://github.com/astral-sh/uv", rev = "855c1917e1e0e2b48c38de71bebc845af016afae" }
 
 [patch.crates-io]
 reqwest-middleware = { git = "https://github.com/astral-sh/reqwest-middleware", rev = "5e3eaf254b5bd481c75d2710eed055f95b756913", features = ["multipart"] }
