+++ py/private/toolchain/BUILD.bazel
@@ -6,23 +6,6 @@ toolchain_type(
     visibility = ["//visibility:public"],
 )
 
-toolchain_type(
-    name = "venv_toolchain_type",
-    visibility = ["//visibility:public"],
-)
-
-# Exec-configured variant: used for build actions that run the venv binary
-# on the exec host (e.g. creating the venv directory in py_venv).
-toolchain_type(
-    name = "venv_exec_toolchain_type",
-    visibility = ["//visibility:public"],
-)
-
-toolchain_type(
-    name = "shim_toolchain_type",
-    visibility = ["//visibility:public"],
-)
-
 # Dummy toolchain backing target — no binary, used as a sentinel.
 # Public so the generated @rules_py_tools repo can reference it for
 # native_build toolchain entries.
