--- py/tools/py/BUILD.bazel
@@ -1,38 +1,0 @@
-load("//bazel/rust:defs.bzl", "rust_library")
-
-rust_library(
-    name = "py",
-    srcs = [
-        "src/lib.rs",
-        "src/pth.rs",
-        "src/unpack.rs",
-        "src/venv.rs",
-    ],
-    compile_data = [
-        "src/__virtualenv.py",
-        "src/activate.tmpl",
-        "src/pyvenv.cfg.tmpl",
-        "src/runfiles_interpreter.tmpl",
-    ],
-    visibility = [
-        "//py/tools/unpack_bin:__pkg__",
-        "//py/tools/venv_bin:__pkg__",
-    ],
-    deps = [
-        "@aspect_rules_py__crates//:indexmap",
-        "@aspect_rules_py__crates//:itertools",
-        "@aspect_rules_py__crates//:miette",
-        "@aspect_rules_py__crates//:pathdiff",
-        "@aspect_rules_py__crates//:percent-encoding-2.3.1",
-        "@aspect_rules_py__crates//:sha256",
-        "@aspect_rules_py__crates//:tempfile",
-        "@aspect_rules_py__crates//:uv-cache",
-        "@aspect_rules_py__crates//:uv-distribution-filename",
-        "@aspect_rules_py__crates//:uv-extract",
-        "@aspect_rules_py__crates//:uv-install-wheel",
-        "@aspect_rules_py__crates//:uv-pypi-types",
-        "@aspect_rules_py__crates//:uv-python",
-        "@aspect_rules_py__crates//:uv-virtualenv",
-        "@aspect_rules_py__crates//:walkdir",
-    ],
-)
