--- py/tools/py/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "py"
-version.workspace = true
-categories.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-edition.workspace = true
-readme.workspace = true
-rust-version.workspace = true
-
-[features]
-debug = []
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-indexmap = "2.11.4"
-itertools = { workspace = true }
-miette = { workspace = true }
-pathdiff = "0.2.3"
-percent-encoding = "2.3.1"
-sha256 = "1.6.0"
-tempfile = { workspace = true }
-uv-cache = { workspace = true }
-uv-distribution-filename = { workspace = true }
-uv-extract = { workspace = true }
-uv-install-wheel = { workspace = true }
-uv-pypi-types = { workspace = true }
-uv-python = { workspace = true }
-uv-virtualenv = { workspace = true }
-walkdir = "2.5.0"
