--- py/tools/py/src/activate.tmpl
@@ -1,78 +1,0 @@
-# Adapted from CPython Lib/venv/scripts/common/activate
-
-deactivate () {
-    # reset old environment variables
-    if [ -n "${_OLD_VIRTUAL_PATH:-}" ] ; then
-        PATH="${_OLD_VIRTUAL_PATH:-}"
-        export PATH
-    fi
-
-    if [ "${_OLD_VIRTUAL_PYTHONHOME:-}" = "_activate_undef" ]; then
-        unset _OLD_VIRTUAL_PYTHONHOME
-        unset PYTHONHOME
-    elif [ -n "${_OLD_VIRTUAL_PYTHONHOME:-}" ] ; then
-        PYTHONHOME="${_OLD_VIRTUAL_PYTHONHOME:-}"
-        export PYTHONHOME
-    fi
-
-    # Call hash to forget past locations. Without forgetting past locations the
-    # $PATH changes we made may not be respected. See "man bash" for more
-    # details. hash is usually a builtin of your shell
-    hash -r 2> /dev/null
-
-    if [ -n "${_OLD_VIRTUAL_PS1:-}" ] ; then
-        PS1="${_OLD_VIRTUAL_PS1:-}"
-        export PS1
-    fi
-
-    unset _OLD_VIRTUAL_PS1
-    unset _OLD_VIRTUAL_PATH
-    unset _OLD_VIRTUAL_PYTHONHOME
-    unset VIRTUAL_ENV
-    unset VIRTUAL_ENV_PROMPT
-
-    # Unset Bazel-injected vars
-{{ENVVARS_UNSET}}
-
-    # Unset vars we set with the runfiles interpreter. Note that this needs to
-    # be conditional so we don't throw this state out under tests or run.
-    if [ "${_OLD_RUNFILES_DIR:-}" = "_activate_undef" ]; then
-        unset RUNFILES_DIR
-        unset RUNFILES_MANIFEST_FILE
-    fi
-
-    if [ ! "${1:-}" = "nondestructive" ] ; then
-    # Self destruct!
-        unset -f deactivate
-    fi
-}
-
-# unset irrelevant variables
-deactivate nondestructive
-
-# For ZSH, emulate BASH_SOURCE.
-# The runfiles library code has some deps on this so we just set it :/
-: "${BASH_SOURCE:=$0}"
-
-VIRTUAL_ENV="$(dirname "$(dirname "${BASH_SOURCE}")")"
-export VIRTUAL_ENV
-
-# unset PYTHONHOME if set
-# this will fail if PYTHONHOME is set to the empty string (which is bad anyway)
-# could use `if (set -u; : $PYTHONHOME) ;` in bash.
-_OLD_VIRTUAL_PYTHONHOME="${PYTHONHOME:-_activate_undef}"
-unset PYTHONHOME
-
-_OLD_VIRTUAL_PATH="$PATH"
-
-# Aspect additions
-# We set these before runfiles initialization so that we can use it as part of a fallback path
-{{ENVVARS}}
-
-# Now we can put the venv's absolute bin on the path
-PATH="$VIRTUAL_ENV/bin:$PATH"
-export PATH
-
-# Call hash to forget past commands. Without forgetting
-# past commands the $PATH changes we made may not be respected
-hash -r 2> /dev/null
