--- py/tools/py/src/lib.rs
@@ -1,8 +1,0 @@
-pub mod pth;
-pub mod unpack;
-pub mod venv;
-
-pub use unpack::unpack_wheel;
-pub use venv::create_venv;
-
-pub use pth::{CollisionResolutionStrategy, PthFile};
