--- py/tools/py/src/pyvenv.cfg.tmpl
@@ -1,7 +1,0 @@
-home = ./bin/
-implementation = CPython
-version_info = {{MAJOR}}.{{MINOR}}.{{PATCH}}
-include-system-site-packages = {{INCLUDE_SYSTEM_SITE}}
-aspect-include-user-site-packages = {{INCLUDE_USER_SITE}}
-relocatable = true
-{{INTERPRETER}}
