--- py/tools/py/src/unpack.rs
@@ -1,92 +1,0 @@
-use std::{
-    fs,
-    path::{Path, PathBuf},
-    str::FromStr,
-};
-
-use itertools::Itertools;
-use miette::{Context, IntoDiagnostic, Result};
-use percent_encoding::percent_decode_str;
-
-const RELOCATABLE_SHEBANG: &'static str = r#"/bin/sh
-'''exec' "$(dirname -- "$(realpath -- "$0")")"/'python3' "$0" "$@"
-' '''
-"#;
-
-pub fn unpack_wheel(
-    version_major: u8,
-    version_minor: u8,
-    location: &Path,
-    wheel: &Path,
-) -> Result<()> {
-    let wheel = if wheel.is_file() {
-        wheel.to_owned()
-    } else {
-        fs::read_dir(wheel)
-            .into_diagnostic()?
-            .filter_map(|res| res.ok())
-            .map(|dir_entry| dir_entry.path())
-            .filter_map(|path| {
-                if path.extension().map_or(false, |ext| ext == "whl") {
-                    Some(path)
-                } else {
-                    None
-                }
-            })
-            .exactly_one()
-            .into_diagnostic()
-            .wrap_err_with(|| "Didn't find exactly one wheel file to install!")?
-    };
-    let wheel_file_reader = fs::File::open(&wheel).into_diagnostic()?;
-
-    let temp = tempfile::tempdir().into_diagnostic()?;
-
-    let _ = uv_extract::unzip(wheel_file_reader, temp.path()).into_diagnostic()?;
-
-    let site_packages_dir = location
-        .join("lib")
-        .join(format!("python{}.{}", version_major, version_minor,))
-        .join("site-packages");
-
-    let scheme = uv_pypi_types::Scheme {
-        purelib: site_packages_dir.to_path_buf(),
-        platlib: site_packages_dir.to_path_buf(),
-        // No windows support.
-        scripts: location.join("bin"),
-        data: site_packages_dir.to_path_buf(),
-        include: location.join("lib").join("include"),
-    };
-
-    let layout = uv_install_wheel::Layout {
-        sys_executable: PathBuf::from(RELOCATABLE_SHEBANG),
-        python_version: (version_major, version_minor),
-        // Don't stamp in the path to the interpreter into the generated bins
-        // as we don't want an absolute path here.
-        // Perhaps this should be set to just "python" so it picks up the one in the venv path?
-        os_name: "".to_string(),
-        scheme,
-    };
-
-    let filename = wheel
-        .file_name()
-        .and_then(|f| f.to_str())
-        .expect("Expected to get filename from wheel path");
-    let filename = percent_decode_str(filename).decode_utf8_lossy();
-    let wheel_file_name =
-        uv_distribution_filename::WheelFilename::from_str(&filename).into_diagnostic()?;
-
-    uv_install_wheel::linker::install_wheel(
-        &layout,
-        false,
-        temp.path(),
-        &wheel_file_name,
-        None,
-        None,
-        Some("aspect_rule_py"),
-        uv_install_wheel::linker::LinkMode::Copy,
-        &uv_install_wheel::linker::Locks::default(),
-    )
-    .into_diagnostic()?;
-
-    Ok(())
-}
