--- py/tools/runfiles/BUILD.bazel
@@ -1,7 +1,0 @@
-load("@rules_rust//rust:defs.bzl", "rust_library")
-
-rust_library(
-    name = "runfiles",
-    srcs = ["src/lib.rs"],
-    visibility = ["//visibility:public"],
-)
