--- py/tools/runfiles/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "runfiles"
-version.workspace = true
-categories.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-edition.workspace = true
-readme.workspace = true
-rust-version.workspace = true
-
-[dependencies]
