--- py/tools/runfiles/README.md
@@ -1,5 +1,0 @@
-# Runfiles
-
-Vendored from rules_rust since we're stuck on an old rules_rust which doesn't
-yet have that target, using the dep via rules_rust breaks Cargo/rust-analyzer
-anyway and the crates.io release is stale.
