+++ py/tools/unpack/BUILD.bazel
@@ -7,24 +7,23 @@ rust_binary(
     srcs = [
         "src/main.rs",
     ],
+    visibility = ["//visibility:public"],
     deps = [
-        "//py/tools/py",
         "@aspect_rules_py__crates//:clap",
+        "@aspect_rules_py__crates//:itertools",
         "@aspect_rules_py__crates//:miette",
+        "@aspect_rules_py__crates//:percent-encoding-2.3.1",
+        "@aspect_rules_py__crates//:tempfile",
+        "@aspect_rules_py__crates//:uv-distribution-filename",
+        "@aspect_rules_py__crates//:uv-extract",
+        "@aspect_rules_py__crates//:uv-install-wheel",
+        "@aspect_rules_py__crates//:uv-pypi-types",
     ],
 )
 
 multi_platform_rust_binaries(
     name = "bins",
     target = ":unpack",
-)
-
-alias(
-    name = "unpack_bin",
-    actual = ":unpack",
-    visibility = [
-        "//visibility:public",
-    ],
 )
 
 release(
