+++ py/tools/unpack/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "venv_shim"
+name = "unpack"
 version.workspace = true
 categories.workspace = true
 homepage.workspace = true
@@ -9,16 +9,19 @@ edition.workspace = true
 readme.workspace = true
 rust-version.workspace = true
 
-[features]
-debug = []
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [[bin]]
-name = "python_shim"
+name = "unpack"
 path = "src/main.rs"
 
 [dependencies]
+clap = { workspace = true }
+itertools = { workspace = true }
 miette = { workspace = true }
-runfiles = { path = "../runfiles" }
-which = "6.0.3"
+percent-encoding = "2.3.1"
+tempfile = { workspace = true }
+uv-distribution-filename = { workspace = true }
+uv-extract = { workspace = true }
+uv-install-wheel = { workspace = true }
+uv-pypi-types = { workspace = true }
