--- py/tools/venv_bin/BUILD.bazel
@@ -1,33 +1,0 @@
-load("//bazel/release:release.bzl", "release")
-load("//bazel/rust:defs.bzl", "rust_binary")
-load("//bazel/rust:multi_platform_rust_binaries.bzl", "multi_platform_rust_binaries")
-
-rust_binary(
-    name = "venv",
-    srcs = [
-        "src/main.rs",
-    ],
-    deps = [
-        "//py/tools/py",
-        "@aspect_rules_py__crates//:clap",
-        "@aspect_rules_py__crates//:miette",
-    ],
-)
-
-multi_platform_rust_binaries(
-    name = "bins",
-    target = ":venv",
-)
-
-alias(
-    name = "venv_bin",
-    actual = ":venv",
-    visibility = [
-        "//visibility:public",
-    ],
-)
-
-release(
-    name = "release",
-    targets = [":bins"],
-)
