--- py/tools/venv_bin/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "venv_bin"
-version.workspace = true
-categories.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-edition.workspace = true
-readme.workspace = true
-rust-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "venv"
-path = "src/main.rs"
-
-[dependencies]
-clap = { workspace = true }
-miette = { workspace = true }
-py = { path = "../py" }
