--- py/tools/venv_shim/BUILD.bazel
@@ -1,33 +1,0 @@
-load("//bazel/release:release.bzl", "release")
-load("//bazel/rust:defs.bzl", "rust_binary")
-load("//bazel/rust:multi_platform_rust_binaries.bzl", "multi_platform_rust_binaries")
-
-rust_binary(
-    name = "shim",
-    srcs = [
-        "src/main.rs",
-    ],
-    deps = [
-        "//py/tools/runfiles",
-        "@aspect_rules_py__crates//:miette",
-        "@aspect_rules_py__crates//:which",
-    ],
-)
-
-multi_platform_rust_binaries(
-    name = "bins",
-    target = ":shim",
-)
-
-alias(
-    name = "venv_shim",
-    actual = ":shim",
-    visibility = [
-        "//visibility:public",
-    ],
-)
-
-release(
-    name = "release",
-    targets = [":bins"],
-)
